move DowngradeParameterTypeWideningRector to more precise ClassMethod scope
