+++ rules-tests/DowngradePhp72/Rector/ClassMethod/DowngradeParameterTypeWideningRector/Fixture/keep_interface_downgraded.php.inc
@@ -1,0 +1,12 @@
+<?php
+
+namespace Rector\Tests\DowngradePhp72\Rector\ClassMethod\DowngradeParameterTypeWideningRector\Fixture;
+
+use Rector\Tests\DowngradePhp72\Rector\ClassMethod\DowngradeParameterTypeWideningRector\Source\SomeExternalContainerInterface;
+
+class KeepInterfaceDonwgraded implements SomeExternalContainerInterface
+{
+    public function get($id)
+    {
+    }
+}
