+++ rules-tests/DowngradePhp72/Rector/ClassMethod/DowngradeParameterTypeWideningRector/Fixture/keep_trait_and_interface_same.php.inc
@@ -1,0 +1,20 @@
+<?php
+
+namespace Rector\Tests\DowngradePhp72\Rector\ClassMethod\DowngradeParameterTypeWideningRector\Fixture;
+
+final class KeepTraitAndInterfaceSame implements AnyContainerInterface
+{
+    use ServiceLocatorTrait;
+}
+
+trait ServiceLocatorTrait
+{
+    public function has(string $name)
+    {
+    }
+}
+
+interface AnyContainerInterface
+{
+    public function has(string $name);
+}
