+++ rules-tests/DowngradePhp72/Rector/ClassMethod/DowngradeParameterTypeWideningRector/Fixture/second_position.php.inc
@@ -1,0 +1,38 @@
+<?php
+
+namespace Rector\Tests\DowngradePhp72\Rector\ClassMethod\DowngradeParameterTypeWideningRector\Fixture;
+
+interface SecondPositionInterface
+{
+    public function has(string $id, callable $callable);
+}
+
+final class SomeSecondPosition implements SecondPositionInterface
+{
+    public function has(string $id, $callable)
+    {
+    }
+}
+
+?>
+-----
+<?php
+
+namespace Rector\Tests\DowngradePhp72\Rector\ClassMethod\DowngradeParameterTypeWideningRector\Fixture;
+
+interface SecondPositionInterface
+{
+    /**
+     * @param callable $callable
+     */
+    public function has(string $id, $callable);
+}
+
+final class SomeSecondPosition implements SecondPositionInterface
+{
+    public function has(string $id, $callable)
+    {
+    }
+}
+
+?>
