+++ rules-tests/DowngradePhp72/Rector/ClassMethod/DowngradeParameterTypeWideningRector/Fixture/skip_nullable_string_from_external.php.inc
@@ -1,0 +1,14 @@
+<?php
+
+namespace Rector\Tests\DowngradePhp72\Rector\ClassMethod\DowngradeParameterTypeWideningRector\Fixture;
+
+use Rector\Tests\DowngradePhp72\Rector\ClassMethod\DowngradeParameterTypeWideningRector\Source\ParentNullableString;
+
+final class SkipNullableStringFromExternal extends ParentNullableString
+{
+    public function load(string $value = null)
+    {
+    }
+}
+
+?>
