+++ rules-tests/DowngradePhp72/Rector/ClassMethod/DowngradeParameterTypeWideningRector/Fixture/skip_nullable_string_from_external_trait.php.inc
@@ -1,0 +1,14 @@
+<?php
+
+namespace Rector\Tests\DowngradePhp72\Rector\ClassMethod\DowngradeParameterTypeWideningRector\Fixture;
+
+use Rector\Tests\DowngradePhp72\Rector\ClassMethod\DowngradeParameterTypeWideningRector\Source\vendor\NullableStringTrait;
+
+final class SkipNullableStringFromExternalTrait
+{
+    use NullableStringTrait;
+
+    public function load(string $value = null)
+    {
+    }
+}
