+++ rules-tests/DowngradePhp72/Rector/ClassMethod/DowngradeParameterTypeWideningRector/Fixture/skip_required_removal.php.inc
@@ -1,0 +1,14 @@
+<?php
+
+namespace Rector\Tests\DowngradePhp72\Rector\ClassMethod\DowngradeParameterTypeWideningRector\Fixture;
+
+use PhpParser\Node\Param;
+use PHPStan\Type\Type;
+use Rector\TypeDeclaration\Contract\TypeInferer\ParamTypeInfererInterface;
+
+final class SkipRequiredRemoval implements ParamTypeInfererInterface
+{
+    public function inferParam(Param $param): Type
+    {
+    }
+}
