+++ rules-tests/DowngradePhp72/Rector/ClassMethod/DowngradeParameterTypeWideningRector/Source/AnotherContainerInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rector\Tests\DowngradePhp72\Rector\ClassMethod\DowngradeParameterTypeWideningRector\Source;
+
+interface AnotherContainerInterface
+{
+    public function get($name);
+}
