--- rules-tests/DowngradePhp72/Rector/Class_/DowngradeParameterTypeWideningRector/Fixture/skip_extra_parameter_container_interface.php.inc
@@ -1,12 +1,0 @@
-<?php
-
-namespace Rector\Tests\DowngradePhp72\Rector\Class_\DowngradeParameterTypeWideningRector\Fixture;
-
-use Psr\Container\ContainerInterface;
-
-interface SkipExtraParameterContainerInterface extends ContainerInterface
-{
-    public function get(string $id, int $invalidBehavior = 100);
-
-    public function has(string $id);
-}
