--- rules-tests/DowngradePhp72/Rector/Class_/DowngradeParameterTypeWideningRector/Fixture/skip_nothing_happens.php.inc
@@ -1,15 +1,0 @@
-<?php
-
-namespace Rector\Tests\DowngradePhp72\Rector\Class_\DowngradeParameterTypeWideningRector\Fixture;
-
-interface KeepOriginalIfSharedType
-{
-    public function test(array $input);
-}
-
-final class UpdateBoth implements KeepOriginalIfSharedType
-{
-    public function test(array $input)
-    {
-    }
-}
