--- rules-tests/DowngradePhp72/Rector/Class_/DowngradeParameterTypeWideningRector/Fixture/skip_nullable_string_from_external_trait.php.inc
@@ -1,14 +1,0 @@
-<?php
-
-namespace Rector\Tests\DowngradePhp72\Rector\Class_\DowngradeParameterTypeWideningRector\Fixture;
-
-use Rector\Tests\DowngradePhp72\Rector\Class_\DowngradeParameterTypeWideningRector\Source\vendor\NullableStringTrait;
-
-final class SkipNullableStringFromExternalTrait
-{
-    use NullableStringTrait;
-
-    public function load(string $value = null)
-    {
-    }
-}
