--- rules-tests/DowngradePhp72/Rector/Class_/DowngradeParameterTypeWideningRector/Source/ParentNullableString.php
@@ -1,12 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Rector\Tests\DowngradePhp72\Rector\Class_\DowngradeParameterTypeWideningRector\Source;
-
-class ParentNullableString
-{
-    public function load(string $value = null)
-    {
-    }
-}
