--- rules-tests/DowngradePhp72/Rector/Class_/DowngradeParameterTypeWideningRector/Source/SomeExternalContainerInterface.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Rector\Tests\DowngradePhp72\Rector\Class_\DowngradeParameterTypeWideningRector\Source;
-
-interface SomeExternalContainerInterface
-{
-    public function get(string $id);
-}
